+++ v1-to-v2-data-conversion/v1-to-v2-data-conversion/Data/v2/ServiceType.cs
@@ -10,7 +10,7 @@ namespace Conversion.Data.v2
             Project = new HashSet<Project>();
         }
 
-        public long ServiceTypeId { get; set; }
+        public int ServiceTypeId { get; set; }
         public string ServiceTypeValue { get; set; }
 
         public ICollection<Project> Project { get; set; }
